+++ index.css
@@ -45,7 +45,7 @@
   justify-content: flex-end;
 }
 
-header {
+.header {
   text-align: center;
   margin-top: 50px;
   margin-bottom: 30px;
