+++ index.html
@@ -12,9 +12,9 @@
   <nav class="nav-menu">
     <h3 class="nav-header">Awesome Books</h3>
     <div class="nav-buttons">
-      <button  type="button" class="button-list">List</button>
+      <button type="button" class="button-list">List</button>
       <div class="separator"></div>
-      <button  type="button" class="button-add">Add new</button>
+      <button type="button" class="button-add">Add new</button>
       <div class="separator"></div>
       <button type="button" class="button-contact">Contact</button>
     </div>
@@ -23,16 +23,14 @@
     <p class="date-time"></p>
   </div>
   <header>
-    <h1 class="header-list">All awesome books</h1>
-    <h1 class="header-add">Add a new book</h1>
-    <h1 class="header-contact">Contact information</h1>
+    <h1 class="header"></h1>
   </header>
   <div class="book-list">
   </div>
   <form class="form" action="#" method="post">
     <input name="title" placeholder="Title" required /><br><br>
     <input name="author" placeholder="Author" required /><br><br>
-    <button  class="form-button" type="submit">Add</button>
+    <button class="form-button" type="submit">Add</button>
   </form>
   <section class="contact-section">
     <p>Do you have any questions or you just want to say "Hello"?</p>
@@ -46,9 +44,8 @@
   <footer class="footer">
     <p>Copyright</p>
   </footer>
-  <script src="./index.js"></script>
+  <script type="module" src="./index.js"></script>
   <script src="./modules/date.js"></script>
-  <script src="./modules/pages.js"></script>
 </body>
 
 </html>
